Update test codes

+++ tests/test_configuration.py
@@ -3,7 +3,6 @@ Tests of the 'Configuration' step
 """
 import random
 from pathlib import Path
-
 import pytest
 from cwas.configuration import Configuration
 from cwas.env import Env
@@ -14,7 +13,6 @@ import sys
 @pytest.fixture(scope="module")
 def vep_mock(cwas_workspace: Path):
     return cwas_workspace / "vep"
-
 
 @pytest.fixture(scope="module", autouse=True)
 def setup(cwas_workspace: Path, vep_mock: Path):
@@ -28,8 +26,6 @@ def create_vep_dir(
     vep_dir, vep_conserv, vep_loftee, vep_msdb, vep_ances, vep_gerp
 ):
     vep_dir.mkdir()
-    #create_vep_conserv(vep_conserv)
-    #create_misdb(vep_msdb)
     Path(vep_conserv).touch()
     Path(vep_msdb).touch()
     Path(vep_loftee).mkdir()
@@ -53,7 +49,6 @@ def teardown(cwas_workspace: Path, vep_mock: Path):
     for f in cwas_workspace.glob("*"):
         f.unlink()
     cwas_workspace.rmdir()
-
 
 def set_env(cwas_workspace: Path):
     env = Env()
@@ -98,24 +93,19 @@ def invalid_file_path(cwas_workspace: Path):
     filename = f"invalid-{random.randint(1, 1000000)}"
     return cwas_workspace / filename
 
-
 @pytest.fixture
 def create_cwas_config_file(cwas_workspace, cwas_config):
     _create_cwas_config_file(cwas_workspace, cwas_config)
 
+# @pytest.fixture
+# def create_incomplete_cwas_config_file(cwas_workspace, cwas_config):
+#     _unset_required_config(cwas_config)
+#     _create_cwas_config_file(cwas_workspace, cwas_config)
 
-@pytest.fixture
-def create_incomplete_cwas_config_file(cwas_workspace, cwas_config):
-    _unset_required_config(cwas_config)
-    _create_cwas_config_file(cwas_workspace, cwas_config)
-
-
-#optional?
-#@pytest.fixture
-#def create_cwas_config_file_without_optional(cwas_workspace, cwas_config):
-#    _unset_optional_config(cwas_config)
-#    _create_cwas_config_file(cwas_workspace, cwas_config)
-
+# @pytest.fixture
+# def create_cwas_config_file_without_optional(cwas_workspace, cwas_config):
+#     _unset_optional_config(cwas_config)
+#     _create_cwas_config_file(cwas_workspace, cwas_config)
 
 @pytest.fixture
 def create_cwas_config_file_invalid_file_path(
@@ -123,7 +113,6 @@ def create_cwas_config_file_invalid_file_path(
 ):
     _set_invalid_file_path(cwas_config, invalid_file_path)
     _create_cwas_config_file(cwas_workspace, cwas_config)
-
 
 @pytest.fixture
 def create_cwas_config_file_invalid_dir_path(
@@ -140,10 +129,8 @@ def create_cwas_config_file_invalid_vep_path(
     _set_invalid_vep_path(cwas_config, invalid_file_path)
     _create_cwas_config_file(cwas_workspace, cwas_config)
 
-
 def _create_cwas_config_file(cwas_workspace, cwas_config):
     config_path = cwas_workspace / "configuration.txt"
-
     with config_path.open("w") as config_file:
         for k, v in cwas_config.items():
             print(f"{k}={str(v)}", file=config_file)
@@ -157,13 +144,8 @@ def _unset_required_config(cwas_config):
     )
     cwas_config[random_config_key] = ""
 
-
-def _unset_optional_config(cwas_config):
-    random_config_key = random.choice(
-        ["ANNOTATION_KEY_CONFIG"]
-    )
-    cwas_config[random_config_key] = ""
-
+# def _unset_optional_config(cwas_config):
+#     cwas_config["ANNOTATION_KEY_CONFIG"] = None  # Set to None to simulate the absence of optional config
 
 def _set_invalid_file_path(cwas_config, invalid_file_path):
     random_file_key = random.choice(
@@ -171,71 +153,61 @@ def _set_invalid_file_path(cwas_config, invalid_file_path):
     )
     cwas_config[random_file_key] = invalid_file_path
 
-
 def _set_invalid_dir_path(cwas_config, invalid_dir_path):
     cwas_config["ANNOTATION_DATA_DIR"] = invalid_dir_path
 
-
 def _set_invalid_vep_path(cwas_config, invalid_vep_path):
     cwas_config["VEP"] = invalid_vep_path
-
 
 @pytest.fixture
 def configuration_inst():
     sys.argv = ['cwas', 'configuration']
     inst = cwas.cli.main()
-    #inst = Configuration.get_instance()
     return inst
 
-
 def test_run_configuration_with_incomplete(
-    configuration_inst, create_incomplete_cwas_config_file,
+    configuration_inst, cwas_workspace, cwas_config
 ):
+    _unset_required_config(cwas_config)
+    _create_cwas_config_file(cwas_workspace, cwas_config)
     with pytest.raises(ValueError):
         configuration_inst.run()
 
-
 def test_run_configuration(
-    cwas_workspace, configuration_inst, create_cwas_config_file
+    cwas_workspace, configuration_inst, cwas_config
 ):
+    _create_cwas_config_file(cwas_workspace, cwas_config)
     configuration_inst.run()
     _check_config_outputs(cwas_workspace)
-
 
 def test_run_configuration_without_optional(
-    cwas_workspace, configuration_inst, create_cwas_config_file_without_optional
+    cwas_workspace, configuration_inst
 ):
     configuration_inst.run()
-    _check_config_outputs(cwas_workspace)
+    _check_config_outputs(cwas_workspace, optional=False)
 
-
-def _check_config_outputs(cwas_workspace):
+def _check_config_outputs(cwas_workspace, optional=True):
     data_dir_symlink = cwas_workspace / "annotation-data"
     gene_matrix_symlink = cwas_workspace / "gene_matrix.txt"
-    bed_key_list = cwas_workspace / "annotation_keys.yaml"
     category_domain_list = cwas_workspace / "category_domain.yaml"
     redundant_category_table = cwas_workspace / "redundant_category.txt"
 
     assert data_dir_symlink.is_dir() and data_dir_symlink.is_symlink()
     assert gene_matrix_symlink.is_file() and gene_matrix_symlink.is_symlink()
-    assert bed_key_list.is_file()
     assert category_domain_list.is_file()
     assert redundant_category_table.is_file()
+
+    if optional:
+        bed_key_list = cwas_workspace / "annotation_keys.yaml"
+        assert bed_key_list.is_file()
 
     # Teardown
     data_dir_symlink.unlink()
     gene_matrix_symlink.unlink()
-    bed_key_list.unlink()
     category_domain_list.unlink()
     redundant_category_table.unlink()
-
-
-def test_run_configuration_with_incomplete(
-    configuration_inst, create_incomplete_cwas_config_file,
-):
-    with pytest.raises(ValueError):
-        configuration_inst.run()
-
+    if optional:
+        bed_key_list.unlink()
 
 def test_run_configuration_with_invalid_file_path(
     configuration_inst, create_cwas_config_file_invalid_file_path
@@ -243,20 +215,17 @@ def test_run_configuration_with_invalid_file_path(
     with pytest.raises(FileNotFoundError):
         configuration_inst.run()
 
-
 def test_run_configuration_with_invalid_dir_path(
     configuration_inst, create_cwas_config_file_invalid_dir_path
 ):
     with pytest.raises(NotADirectoryError):
         configuration_inst.run()
 
-
 def test_run_configuration_with_invalid_vep_path(
     configuration_inst, create_cwas_config_file_invalid_vep_path
 ):
     with pytest.raises(ValueError):
         configuration_inst.run()
-
 
 def test_env_after_run_configuration(
     configuration_inst, create_cwas_config_file
@@ -283,16 +252,14 @@ def test_env_after_run_configuration(
     for env_key in env_keys:
         assert configuration_inst.get_env(env_key)
 
-
 def test_get_inst_without_env():
     _make_env_empty()
     with pytest.raises(RuntimeError):
         sys.argv = ['cwas', 'configuration']
         cwas.cli.main()
-        #Configuration.get_instance()
-
 
 def _make_env_empty():
     env = Env()
     env.reset()
     env.save()
+
